chore(design-tokens): rename global navigation
