+++ packages/react/src/themes/tokens/component/global-navigation-tokens.ts
